Update css for map and locations, add VRC
Add the Visual Resources Center at Fayerweather Hall to the map and list
of locations. Still have to do this by hand, by editing the
_data/makergrounds.csv, _data/tool-categories.json, and
map-geojson/makergrounds-map-data.geojson files.

Added some margin between the locations so it's much easier to tell
which image the title goes with. Before, it was hard to tell if the
title went with the image above or below it.

Also updated the CSS for the map so that the icons don't overlap on
smaller screens when they wrap onto more than one line.